+++ packages/core/hardhat.config.js
@@ -21,12 +21,22 @@ const settings = {
   },
 };
 
+const proxyCompiler = {
+  version: require('./src/solidity-version.json'),
+  settings: {
+    ...settings,
+    evmVersion: 'paris',
+  },
+};
+
 function getNamespacedOverrides() {
   const contracts = fs.readdirSync(path.join(__dirname, 'contracts', 'test'));
   const namespacedContracts = contracts.filter(c => c.startsWith('Namespaced'));
   const overrides = {};
   for (const c of namespacedContracts) {
-    if (c !== 'NamespacedToModify07.sol') {
+    if (c === 'NamespacedToModify07.sol') {
+      overrides[`contracts/test/${c}`] = { version: '0.7.6', settings };
+    } else {
       overrides[`contracts/test/${c}`] = { version: '0.8.20', settings };
     }
   }
@@ -49,6 +59,7 @@ module.exports = {
       { version: '0.7.6', settings },
       { version: '0.8.8', settings },
       { version: '0.8.9', settings },
+      proxyCompiler,
     ],
     overrides: getNamespacedOverrides(),
   },