+++ packages/core/scripts/copy-artifacts.sh
@@ -4,10 +4,12 @@ set -euo pipefail
 
 # Copies proxy artifacts to their location in previous versions for backwards compatibility
 
-mkdir artifacts/contracts/proxy/{,AdminUpgradeabilityProxy.sol,ProxyAdmin.sol}
+mkdir -p artifacts
 
-cp artifacts/contracts/import.sol/AdminUpgradeabilityProxy.json artifacts
-cp artifacts/contracts/import.sol/AdminUpgradeabilityProxy.json artifacts/contracts/proxy/AdminUpgradeabilityProxy.sol
+# Assert that a previous version of @openzeppelin/upgrades-core is installed, which contains legacy artifacts
+if ! grep -q '"version": "1.31.1"' ../../node_modules/@openzeppelin/upgrades-core-legacy/package.json; then
+  echo "Error: @openzeppelin/upgrades-core must depend on a previous version of itself at version 1.31.1"
+  exit 1
+fi
 
-cp artifacts/@openzeppelin/contracts/proxy/transparent/ProxyAdmin.sol/ProxyAdmin.json artifacts
-cp artifacts/@openzeppelin/contracts/proxy/transparent/ProxyAdmin.sol/ProxyAdmin.json artifacts/contracts/proxy/ProxyAdmin.sol
+cp -R ../../node_modules/@openzeppelin/upgrades-core-legacy/artifacts/ artifacts