+++ packages/core/src/standalone.test.ts
@@ -1,6 +1,8 @@
 import _test, { TestFn } from 'ava';
 import { artifacts } from 'hardhat';
 
+import SOLIDITY_VERSION from './solidity-version.json';
+
 import { SolcInput, SolcOutput } from './solc-api';
 import { UpgradeableContract } from './standalone';
 
@@ -21,7 +23,7 @@ test.before(async t => {
 });
 
 test('reports unsafe operation', t => {
-  const impl = new UpgradeableContract('StandaloneV1', t.context.solcInput, t.context.solcOutput);
+  const impl = new UpgradeableContract('StandaloneV1', t.context.solcInput, t.context.solcOutput, {}, SOLIDITY_VERSION);
   const report = impl.getErrorReport();
   t.false(report.ok);
   t.true(report.errors[0].kind === 'delegatecall');
@@ -32,6 +34,8 @@ test('reports unsafe operation - fully qualified name', t => {
     'contracts/test/Standalone.sol:StandaloneV1',
     t.context.solcInput,
     t.context.solcOutput,
+    {},
+    SOLIDITY_VERSION,
   );
   const report = impl.getErrorReport();
   t.false(report.ok);
@@ -39,25 +43,55 @@ test('reports unsafe operation - fully qualified name', t => {
 });
 
 test('reports storage upgrade errors', t => {
-  const v1 = new UpgradeableContract('StandaloneV1', t.context.solcInput, t.context.solcOutput);
+  const v1 = new UpgradeableContract('StandaloneV1', t.context.solcInput, t.context.solcOutput, {}, SOLIDITY_VERSION);
 
-  const v2Good = new UpgradeableContract('StandaloneV2Good', t.context.solcInput, t.context.solcOutput);
+  const v2Good = new UpgradeableContract(
+    'StandaloneV2Good',
+    t.context.solcInput,
+    t.context.solcOutput,
+    {},
+    SOLIDITY_VERSION,
+  );
   const goodReport = v1.getStorageUpgradeReport(v2Good);
   t.true(goodReport.ok);
 
-  const v2Bad = new UpgradeableContract('StandaloneV2Bad', t.context.solcInput, t.context.solcOutput);
+  const v2Bad = new UpgradeableContract(
+    'StandaloneV2Bad',
+    t.context.solcInput,
+    t.context.solcOutput,
+    {},
+    SOLIDITY_VERSION,
+  );
   const badReport = v1.getStorageUpgradeReport(v2Bad);
   t.false(badReport.ok);
 });
 
 test('dont report renamed version update', t => {
-  const v1 = new UpgradeableContract('StandaloneRenameV1', t.context.solcInput, t.context.solcOutput);
+  const v1 = new UpgradeableContract(
+    'StandaloneRenameV1',
+    t.context.solcInput,
+    t.context.solcOutput,
+    {},
+    SOLIDITY_VERSION,
+  );
 
-  const v2 = new UpgradeableContract('StandaloneRenameV2', t.context.solcInput, t.context.solcOutput);
+  const v2 = new UpgradeableContract(
+    'StandaloneRenameV2',
+    t.context.solcInput,
+    t.context.solcOutput,
+    {},
+    SOLIDITY_VERSION,
+  );
   const goodReport = v1.getStorageUpgradeReport(v2);
   t.true(goodReport.ok);
 
-  const v3 = new UpgradeableContract('StandaloneRenameV3', t.context.solcInput, t.context.solcOutput);
+  const v3 = new UpgradeableContract(
+    'StandaloneRenameV3',
+    t.context.solcInput,
+    t.context.solcOutput,
+    {},
+    SOLIDITY_VERSION,
+  );
   const goodReport2 = v2.getStorageUpgradeReport(v3);
   t.true(goodReport2.ok);
 });