+++ packages/plugin-hardhat/contracts/GreeterV3.sol
@@ -28,3 +28,6 @@ contract GreeterV3 {
 
 import "./utils/Proxiable.sol";
 contract GreeterV3Proxiable is GreeterV3, Proxiable {}
+
+import "./utils/Proxiable40.sol";
+contract GreeterV3Proxiable40 is GreeterV3, Proxiable40 {}