+++ packages/plugin-hardhat/contracts/utils/Proxiable40.sol
@@ -2,10 +2,12 @@ pragma solidity >= 0.4.22 <0.8.0;
 
 // This contract is for testing only, it is not safe for use in production.
 
-contract Proxiable50 {
+contract Proxiable40 {
     bytes32 internal constant _IMPLEMENTATION_SLOT = 0x360894a13ba1a3210667c828492db98dca3e2076cc3735a920a3ca505d382bbc;
 
-    string public constant UPGRADE_INTERFACE_VERSION = "5.0.0";
+    function upgradeTo(address newImplementation) external {
+        _setImplementation(newImplementation);
+    }
 
     function upgradeToAndCall(address newImplementation, bytes calldata data) external {
         _setImplementation(newImplementation);
@@ -25,14 +27,6 @@ contract Proxiable50 {
              */
             (bool success, ) = address(this).call(data);
             require(success, "upgrade call reverted");
-        } else {
-            _checkNonPayable();
-        }
-    }
-
-    function _checkNonPayable() private {
-        if (msg.value > 0) {
-            revert('non-payable upgrade call');
         }
     }