+++ packages/plugin-hardhat/src/deploy-beacon.ts
@@ -6,6 +6,7 @@ import { Deployment } from '@openzeppelin/upgrades-core';
 import { DeployBeaconOptions, deploy, DeployTransaction, getUpgradeableBeaconFactory, deployBeaconImpl } from './utils';
 import { disableDefender } from './defender/utils';
 import { attach, getSigner } from './utils/ethers';
+import { getInitialOwner } from './utils/initial-owner';
 
 export interface DeployBeaconFunction {
   (ImplFactory: ContractFactory, opts?: DeployBeaconOptions): Promise<Contract>;
@@ -17,8 +18,18 @@ export function makeDeployBeacon(hre: HardhatRuntimeEnvironment, defenderModule:
 
     const { impl } = await deployBeaconImpl(hre, ImplFactory, opts);
 
-    const UpgradeableBeaconFactory = await getUpgradeableBeaconFactory(hre, getSigner(ImplFactory.runner));
-    const beaconDeployment: Deployment & DeployTransaction = await deploy(hre, opts, UpgradeableBeaconFactory, impl);
+    const signer = getSigner(ImplFactory.runner);
+    const UpgradeableBeaconFactory = await getUpgradeableBeaconFactory(hre, signer);
+
+    const initialOwner = await getInitialOwner(opts, signer);
+
+    const beaconDeployment: Deployment & DeployTransaction = await deploy(
+      hre,
+      opts,
+      UpgradeableBeaconFactory,
+      impl,
+      initialOwner,
+    );
     const beaconContract = attach(UpgradeableBeaconFactory, beaconDeployment.address);
 
     // @ts-ignore Won't be readonly because beaconContract was created through attach.