+++ packages/plugin-hardhat/src/utils/simulate-deploy.ts
@@ -1,4 +1,4 @@
-import { fetchOrDeploy, fetchOrDeployAdmin, logWarning } from '@openzeppelin/upgrades-core';
+import { fetchOrDeploy } from '@openzeppelin/upgrades-core';
 import type { ContractFactory } from 'ethers';
 import type { HardhatRuntimeEnvironment } from 'hardhat/types';
 import { getDeployData } from './deploy-impl';
@@ -8,25 +8,6 @@ import { UpgradeOptions } from './options';
 // a deployment and record it in the network file. We are able to do this by "simulating" a deployment:
 // for the "deploy" part we pass a function that simply returns the contract to be imported, rather than
 // actually deploying something.
-
-export async function simulateDeployAdmin(
-  hre: HardhatRuntimeEnvironment,
-  ProxyAdminFactory: ContractFactory,
-  opts: UpgradeOptions,
-  adminAddress: string,
-) {
-  const { deployData, simulateDeploy } = await getSimulatedData(hre, ProxyAdminFactory, opts, adminAddress);
-  const manifestAdminAddress = await fetchOrDeployAdmin(deployData.provider, simulateDeploy, opts);
-  if (adminAddress !== manifestAdminAddress) {
-    logWarning(
-      `Imported proxy with admin at '${adminAddress}' which differs from previously deployed admin '${manifestAdminAddress}'`,
-      [
-        `The imported proxy admin is different from the proxy admin that was previously deployed on this network.`,
-        `New proxy deployments will continue to use the admin '${manifestAdminAddress}'.`,
-      ],
-    );
-  }
-}
 
 export async function simulateDeployImpl(
   hre: HardhatRuntimeEnvironment,