+++ packages/plugin-hardhat/test/import-v4.js
@@ -2,21 +2,21 @@ const test = require('ava');
 
 const { ethers, upgrades } = require('hardhat');
 
-const ProxyAdmin = require('../artifacts/@openzeppelin/contracts-5.0/proxy/transparent/ProxyAdmin.sol/ProxyAdmin.json');
-const TransparentUpgradableProxy = require('../artifacts/@openzeppelin/contracts-5.0/proxy/transparent/TransparentUpgradeableProxy.sol/TransparentUpgradeableProxy.json');
+const ProxyAdmin = require('@openzeppelin/upgrades-core/artifacts/@openzeppelin/contracts/proxy/transparent/ProxyAdmin.sol/ProxyAdmin.json');
+const TransparentUpgradableProxy = require('@openzeppelin/upgrades-core/artifacts/@openzeppelin/contracts/proxy/transparent/TransparentUpgradeableProxy.sol/TransparentUpgradeableProxy.json');
 
-const ERC1967Proxy = require('../artifacts/@openzeppelin/contracts-5.0/proxy/ERC1967/ERC1967Proxy.sol/ERC1967Proxy.json');
+const ERC1967Proxy = require('@openzeppelin/upgrades-core/artifacts/@openzeppelin/contracts/proxy/ERC1967/ERC1967Proxy.sol/ERC1967Proxy.json');
 
-const BeaconProxy = require('../artifacts/@openzeppelin/contracts-5.0/proxy/beacon/BeaconProxy.sol/BeaconProxy.json');
-const UpgradableBeacon = require('../artifacts/@openzeppelin/contracts-5.0/proxy/beacon/UpgradeableBeacon.sol/UpgradeableBeacon.json');
+const BeaconProxy = require('@openzeppelin/upgrades-core/artifacts/@openzeppelin/contracts/proxy/beacon/BeaconProxy.sol/BeaconProxy.json');
+const UpgradableBeacon = require('@openzeppelin/upgrades-core/artifacts/@openzeppelin/contracts/proxy/beacon/UpgradeableBeacon.sol/UpgradeableBeacon.json');
 
 test.before(async t => {
   t.context.Greeter = await ethers.getContractFactory('Greeter');
   t.context.GreeterV2 = await ethers.getContractFactory('GreeterV2');
   t.context.GreeterV3 = await ethers.getContractFactory('GreeterV3');
-  t.context.GreeterProxiable = await ethers.getContractFactory('Greeter50Proxiable');
-  t.context.GreeterV2Proxiable = await ethers.getContractFactory('Greeter50V2Proxiable');
-  t.context.GreeterV3Proxiable = await ethers.getContractFactory('GreeterV3Proxiable');
+  t.context.GreeterProxiable = await ethers.getContractFactory('GreeterProxiable40');
+  t.context.GreeterV2Proxiable = await ethers.getContractFactory('GreeterV2Proxiable40');
+  t.context.GreeterV3Proxiable = await ethers.getContractFactory('GreeterV3Proxiable40');
   t.context.CustomProxy = await ethers.getContractFactory('CustomProxy');
   t.context.CustomProxyWithAdmin = await ethers.getContractFactory('CustomProxyWithAdmin');
 
@@ -68,15 +68,13 @@ test('no contract', async t => {
 test('transparent happy path', async t => {
   const { Greeter, GreeterV2, ProxyAdmin, TransparentUpgradableProxy } = t.context;
 
-  const owner = (await ethers.getSigners())[0];
-
   const impl = await Greeter.deploy();
   await impl.waitForDeployment();
-  const admin = await ProxyAdmin.deploy(owner);
+  const admin = await ProxyAdmin.deploy();
   await admin.waitForDeployment();
   const proxy = await TransparentUpgradableProxy.deploy(
     await impl.getAddress(),
-    owner,
+    await admin.getAddress(),
     getInitializerData(Greeter.interface, ['Hello, Hardhat!']),
   );
   await proxy.waitForDeployment();
@@ -115,11 +113,9 @@ test('uups happy path', async t => {
 test('beacon proxy happy path', async t => {
   const { Greeter, GreeterV2, UpgradableBeacon, BeaconProxy } = t.context;
 
-  const owner = (await ethers.getSigners())[0];
-
   const impl = await Greeter.deploy();
   await impl.waitForDeployment();
-  const beacon = await UpgradableBeacon.deploy(await impl.getAddress(), owner);
+  const beacon = await UpgradableBeacon.deploy(await impl.getAddress());
   await beacon.waitForDeployment();
   const proxy = await BeaconProxy.deploy(
     await beacon.getAddress(),
@@ -141,11 +137,9 @@ test('beacon proxy happy path', async t => {
 test('beacon happy path', async t => {
   const { Greeter, GreeterV2, UpgradableBeacon } = t.context;
 
-  const owner = (await ethers.getSigners())[0];
-
   const impl = await Greeter.deploy();
   await impl.waitForDeployment();
-  const beacon = await UpgradableBeacon.deploy(await impl.getAddress(), owner);
+  const beacon = await UpgradableBeacon.deploy(await impl.getAddress());
   await beacon.waitForDeployment();
 
   const beaconImported = await upgrades.forceImport(await beacon.getAddress(), Greeter);
@@ -235,15 +229,13 @@ test('import custom UUPS proxy with admin', async t => {
 test('wrong implementation', async t => {
   const { Greeter, GreeterV2, ProxyAdmin, TransparentUpgradableProxy } = t.context;
 
-  const owner = (await ethers.getSigners())[0];
-
   const impl = await Greeter.deploy();
   await impl.waitForDeployment();
-  const admin = await ProxyAdmin.deploy(owner);
+  const admin = await ProxyAdmin.deploy();
   await admin.waitForDeployment();
   const proxy = await TransparentUpgradableProxy.deploy(
     await impl.getAddress(),
-    owner,
+    await admin.getAddress(),
     getInitializerData(Greeter.interface, ['Hello, Hardhat!']),
   );
   await proxy.waitForDeployment();
@@ -311,26 +303,22 @@ test('same implementation', async t => {
 test('import transparents with different admin', async t => {
   const { Greeter, GreeterV2, ProxyAdmin, TransparentUpgradableProxy } = t.context;
 
-  const owner = (await ethers.getSigners())[0];
-
   const impl = await Greeter.deploy();
   await impl.waitForDeployment();
-  const admin = await ProxyAdmin.deploy(owner);
+  const admin = await ProxyAdmin.deploy();
   await admin.waitForDeployment();
   const proxy = await TransparentUpgradableProxy.deploy(
     await impl.getAddress(),
-    owner,
+    await admin.getAddress(),
     getInitializerData(Greeter.interface, ['Hello, Hardhat!']),
   );
   await proxy.waitForDeployment();
 
-  const owner2 = (await ethers.getSigners())[1];
-
-  const admin2 = await ProxyAdmin.deploy(owner2);
+  const admin2 = await ProxyAdmin.deploy();
   await admin2.waitForDeployment();
   const proxy2 = await TransparentUpgradableProxy.deploy(
     await impl.getAddress(),
-    owner2,
+    await admin2.getAddress(),
     getInitializerData(Greeter.interface, ['Hello, Hardhat 2!']),
   );
   await proxy2.waitForDeployment();
@@ -346,27 +334,4 @@ test('import transparents with different admin', async t => {
   // proxy with a different admin can be imported
   const proxyAddress = await proxy.getAddress();
   await upgrades.upgradeProxy(proxyAddress, GreeterV2);
-});
-
-test('import transparent then upgrade with call', async t => {
-  const { Greeter, GreeterV2, ProxyAdmin, TransparentUpgradableProxy } = t.context;
-
-  const owner = (await ethers.getSigners())[0];
-
-  const impl = await Greeter.deploy();
-  await impl.waitForDeployment();
-  const admin = await ProxyAdmin.deploy(owner);
-  await admin.waitForDeployment();
-  const proxy = await TransparentUpgradableProxy.deploy(
-    await impl.getAddress(),
-    owner,
-    getInitializerData(Greeter.interface, ['Hello, Hardhat!']),
-  );
-  await proxy.waitForDeployment();
-
-  const greeter = await upgrades.forceImport(await proxy.getAddress(), Greeter);
-  t.is(await greeter.greet(), 'Hello, Hardhat!');
-
-  await upgrades.upgradeProxy(greeter, GreeterV2, { call: 'resetGreeting' });
-  t.is(await greeter.greet(), 'Hello World');
 });