+++ packages/plugin-hardhat/test/uups-happy-path-v4.js
@@ -1,0 +1,24 @@
+const test = require('ava');
+
+const { ethers, upgrades } = require('hardhat');
+
+test.before(async t => {
+  t.context.Greeter = await ethers.getContractFactory('GreeterProxiable40');
+  t.context.GreeterV2 = await ethers.getContractFactory('GreeterV2Proxiable40');
+  t.context.GreeterV3 = await ethers.getContractFactory('GreeterV3Proxiable40');
+});
+
+test('happy path', async t => {
+  const { Greeter, GreeterV2, GreeterV3 } = t.context;
+
+  const greeter = await upgrades.deployProxy(Greeter, ['Hello, Hardhat!'], { kind: 'uups' });
+
+  const greeter2 = await upgrades.upgradeProxy(greeter, GreeterV2);
+  await greeter2.waitForDeployment();
+  await greeter2.resetGreeting();
+
+  const greeter3ImplAddr = await upgrades.prepareUpgrade(await greeter.getAddress(), GreeterV3);
+  const greeter3 = GreeterV3.attach(greeter3ImplAddr);
+  const version3 = await greeter3.version();
+  t.is(version3, 'V3');
+});